+++ ini_cfg_parser/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 
 from .ini_cfg_parser import IniParser, DieMode, IniDict, IniItem, IniType, IniValue, IniParserError